requirements.txt
altgraph==0.17.4
attrs==24.2.0
Authlib==0.14.3
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
et-xmlfile==1.1.0
Flask==2.3.3
flask-azure-oauth==0.7.0
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
MouseInfo==0.1.3
msal==1.31.0
numpy==2.1.1
Office365-REST-Python-Client==2.5.12
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pefile==2024.8.26
pillow==10.4.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.1
pywin32-ctypes==0.2.3
pywinstyles==1.8
requests==2.32.3
selenium==4.24.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sv-ttk==2.6.0
tkinterweb==3.24.7
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
